+++ src/model/Model_Budgetsplittransaction.h
@@ -47,7 +47,7 @@ public:
 
 public:
     double get_total(const Data_Set& rows);
-    std::map<int, Data_Set> get_all();
+    std::map<int64, Data_Set> get_all();
     int update(Data_Set& rows, int64 transactionID);
     bool remove(int64 id);
 };
